requirements.txt
appdirs==1.4.4
attrs==20.3.0
CacheControl==0.12.6
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.0
crypto==1.4.1
cryptography==3.4.7
cycler==0.10.0
dataclasses==0.6
distlib==0.3.1
dominate==2.6.0
filelock==3.0.12
firebase==3.0.1
firebase-admin==5.0.0
flake8==3.8.4
flake8-polyfill==1.0.2
Flask==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.3
gcloud==0.18.3
google-api-core==1.26.3
google-api-python-client==2.4.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.1
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.37.1
httplib2==0.19.1
idna==2.10
importlib-metadata==3.3.0
importlib-resources==4.1.1
iniconfig==1.1.1
itsdangerous==2.0.0
Jinja2==3.0.0
jwcrypto==0.8
kiwisolver==1.3.1
MarkupSafe==2.0.0
matplotlib==3.3.4
mccabe==0.6.1
msgpack==1.0.2
Naked==0.1.31
numpy==1.19.5
oauth2client==4.1.3
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
packaging==20.9
pep8-naming==0.11.1
Pillow==8.1.0
pluggy==0.13.1
proto-plus==1.18.1
protobuf==4.0.0rc2
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.10.1
pyflakes==2.2.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.2.2
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.4
python-jwt==3.3.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
scipy==1.5.4
shellescape==3.8.1
six==1.15.0
sseclient==0.0.27
toml==0.10.2
trycourier==1.8.0
twilio==6.58.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
virtualenv==20.2.2
visitor==0.1.3
Werkzeug==2.0.0
WTForms==2.3.3
zipp==3.4.0
